requirements.txt
aiogram>=3.0.0
pydantic-settings
asyncpg
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib